Import progress.

+++ app/src/main/java/org/gnucash/android/db/adapter/TransactionsDbAdapter.java
@@ -116,7 +116,9 @@ public class TransactionsDbAdapter extends DatabaseAdapter<Transaction> {
      */
     @Override
     public void addRecord(@NonNull Transaction transaction, UpdateMethod updateMethod) throws SQLException {
-        Timber.d("Adding transaction to the db via %s", updateMethod.name());
+        Timber.d("%s transaction to the db", updateMethod.name());
+        // Did the transaction have any splits before?
+        final boolean didChange = transaction.id != 0;
         try {
             beginTransaction();
             Split imbalanceSplit = transaction.createAutoBalanceSplit();
@@ -129,8 +131,10 @@ public class TransactionsDbAdapter extends DatabaseAdapter<Transaction> {
             super.addRecord(transaction, updateMethod);
 
             Timber.d("Adding splits for transaction");
-            List<String> splitUIDs = new ArrayList<>(transaction.getSplits().size());
-            for (Split split : transaction.getSplits()) {
+            List<Split> splits = transaction.getSplits();
+            Timber.d("Adding %d splits for transaction", splits.size());
+            List<String> splitUIDs = new ArrayList<>(splits.size());
+            for (Split split : splits) {
                 Timber.d("Replace transaction split in db");
                 if (imbalanceSplit == split) {
                     splitsDbAdapter.addRecord(split, UpdateMethod.insert);
@@ -139,13 +143,15 @@ public class TransactionsDbAdapter extends DatabaseAdapter<Transaction> {
                 }
                 splitUIDs.add(split.getUID());
             }
-            Timber.d("%d splits added", transaction.getSplits().size());
+            Timber.d("%d splits added", splitUIDs.size());
 
-            long deleted = mDb.delete(SplitEntry.TABLE_NAME,
-                SplitEntry.COLUMN_TRANSACTION_UID + " = ? AND "
-                    + SplitEntry.COLUMN_UID + " NOT IN ('" + TextUtils.join("' , '", splitUIDs) + "')",
-                new String[]{transaction.getUID()});
-            Timber.d("%d splits deleted", deleted);
+            if (didChange) {
+                long deleted = mDb.delete(SplitEntry.TABLE_NAME,
+                    SplitEntry.COLUMN_TRANSACTION_UID + " = ? AND "
+                        + SplitEntry.COLUMN_UID + " NOT IN ('" + TextUtils.join("','", splitUIDs) + "')",
+                    new String[]{transaction.getUID()});
+                Timber.d("%d splits deleted", deleted);
+            }
 
             setTransactionSuccessful();
         } finally {
@@ -342,9 +348,8 @@ public class TransactionsDbAdapter extends DatabaseAdapter<Transaction> {
     public Cursor fetchTransactionsModifiedSince(Timestamp timestamp) {
         SQLiteQueryBuilder queryBuilder = new SQLiteQueryBuilder();
         queryBuilder.setTables(TransactionEntry.TABLE_NAME);
-        String startTimeString = Long.toString(timestamp.getTime());
         String where = TransactionEntry.COLUMN_TEMPLATE + "=0 AND " + TransactionEntry.COLUMN_TIMESTAMP + " >= ?";
-        String[] whereArgs = new String[]{startTimeString};
+        String[] whereArgs = new String[]{Long.toString(timestamp.getTime())};
         String orderBy = TransactionEntry.COLUMN_TIMESTAMP + " ASC, " + TransactionEntry.COLUMN_ID + " ASC";
         return queryBuilder.query(mDb, null, where, whereArgs, null, null, orderBy, null);
     }
